Pass configured proxy to build and container boot
This way the build uses the proxy, too. Passing the proxy to the boot
ensures the proxy gets used even if the "apt-get update" step of the
init script is called before the proxy configuration file has been
written. That file is still needed to make the SSH session use the
proxy.